+++ demo/mpx-app/postcss.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: { remove: false },
+    'postcss-rem-to-responsive-pixel': {
+      rootValue: 32,
+      propList: ['*'],
+      transformUnit: 'rpx'
+    }
+  }
+}