+++ demo/mpx-app/src/app.ts
@@ -1,0 +1,10 @@
+import mpx, { createApp } from '@mpxjs/core'
+import apiProxy from '@mpxjs/api-proxy'
+
+mpx.use(apiProxy, { usePromise: true })
+
+createApp({
+  onLaunch () {
+    // onLaunch
+  }
+})