+++ demo/mpx-app/src/index.d.ts
@@ -1,0 +1,1 @@
+// global types declares here