+++ demo/mpx-app/src/pages/index.ts
@@ -1,0 +1,7 @@
+import { createPage } from '@mpxjs/core'
+
+createPage({
+  onLoad () {
+    // onLoad
+  }
+})