+++ demo/mpx-app/tailwind.config.js
@@ -1,0 +1,11 @@
+//tailwind.config.js
+module.exports = {
+  content: ['./src/**/*.{html,js,ts,mpx}'],
+  theme: {
+    extend: {}
+  },
+  plugins: [],
+  corePlugins: {
+    preflight: false
+  }
+}