+++ src/framework/mpx/index.ts
@@ -1,0 +1,8 @@
+import type { UserDefinedOptions } from '@/types'
+import { BaseTemplateWebpackPluginV5 } from '@/base'
+
+export class MpxWeappTailwindcssWebpackPluginV5 extends BaseTemplateWebpackPluginV5 {
+  constructor (options: UserDefinedOptions = {}) {
+    super(options, 'mpx')
+  }
+}