+++ src/framework/v5.ts
@@ -1,2 +1,3 @@
 export * from './rax'
 export * from './native'
+export * from './mpx'