+++ src/framework/vite/index.ts
@@ -1,13 +1,13 @@
 import type { Plugin } from 'vite'
 import { UserDefinedOptions } from '@/types'
-import { getOptions } from '@/shared'
+import { getOptions } from '@/defaults'
 import { templeteHandler } from '@/wxml'
 // import renamePostcssPlugin from '../postcss/plugin'
 // import type { Plugin as PostcssPlugin } from 'postcss'
 // import postcssrc from 'postcss-load-config'
 
 // https://github.com/sonofmagic/weapp-tailwindcss-webpack-plugin/issues/3
-export function ViteWeappTailwindcssPlugin (options: UserDefinedOptions = {}): Plugin {
+export default function ViteWeappTailwindcssPlugin (options: UserDefinedOptions = {}): Plugin {
   const {
     htmlMatcher // cssMatcher, mainCssChunkMatcher
   } = getOptions(options)